Limit attachment collisions between environments
It was possible (though unlikely) that a user could upload an attachment
named, for exmaple, `syllabus.pdf` in staging that could potentially
clobber a production attachment with the same name. For this to happen
it would have to:

* Be attached to a coverage or assignment that has the same type and ID
  as the production attachment.
* Have the same attachment id.

This is pretty unlikely, but we can make it even more unlikely by
replacing the attachment id in the path with the attachment updated_at
timestamp.