User can view and delete attachments
This branch allows the user to view and delete attachments from the
courses page. Clicking on the Attachments link located just below the
outcome description expands and collapses the list of attachments
associated with an assignment. If no attachments belong to an
assignment, the Attachments link will instead read Attach student work.
Clicking on Attach student work will redirect the user to
`assignment#edit`.