Remove old assessments data model
We're now fully migrated to assignments and were given permission by Jon
to go ahead and remove the old data.

I opted for marking the migration as non-reversible because it would
have been non-trivial to write the reversal and reversing the migration
would never bring back dropped data anyway. Realistically, we wouldn't
reverse this migration -- we'd restore the database backup.