View attachments associated with coverages
This branch allows the user to view all attachments associated with a
coverage. By clicking Add a syllabus, the user can attach a file or
files associated with the entire coverage. The expected user behavior is
that each coverage has only one document (a syllabus). From the
expandable view the user can download the current attachments, delete
them, or add another attachment.